test_flexindex
